requirements.txt
arabic-reshaper==2.1.1
asgiref==3.2.10
beautifulsoup4==4.9.3
boto3==1.16.13
botocore==1.19.13
decorator==4.4.2
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-storages==1.10.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
jmespath==0.10.0
Pillow==8.0.1
psycopg2==2.8.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.55
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.11
webencodings==0.5.1
whitenoise==5.2.0
xhtml2pdf==0.2.5
